requirements.txt
customtkinter>=5.2.0
markdown2>=2.4.0
xhtml2pdf>=0.2.17
Pygments>=2.17.0
pywin32>=306
pystray>=0.19.0
Pillow>=10.0.0
tkinterdnd2>=0.4.2
